+++ arch/i2c.c
@@ -269,12 +269,12 @@ void I2C_Update(ARMul_State *state) {
                bit we are going to ignore it; the last bit tells us rNw */
             I2C.LastrNw=I2C.DataBuffer & 1;
             dbug_i2c("I2C simulator got slave address "
-                "%d\n", (int) I2C.DataBuffer);
+                "%u\n", I2C.DataBuffer);
             if ((I2C.DataBuffer & 0xfe)!=0xa0) {
               /* Hey - its a request for a different I2C peripheral - like an A500's timer
                  chip */
               warn_i2c("I2C simulator got wierd slave address "
-                  "%d\n", (int) I2C.DataBuffer);
+                  "%u\n", I2C.DataBuffer);
               I2C.IAmTransmitter=false;
               I2C.NumberOfBitsSoFar=0;
               I2C.state=I2CChipState_Idle;
